move widgets into organized subfolders
