+++ lib/app/layouts/conversation_details/CLAUDE.md
@@ -23,11 +23,12 @@ Displayed as a right-side panel on tablet or pushed screen on mobile.
 
 **Shared Media**
 - `attachment_section_header.dart` — section label + "Show more" action
-- `media_grid_section.dart` — photo/video thumbnail grid
 - `media_gallery_card.dart` — tappable media card → opens `FullscreenMedia`
 - `attachments_loader.dart` — attachment pagination and caching
+- `filters/media_filters_sheet.dart` — shared filters bottom sheet
 
-**Shared Content**
-- `links_section.dart` — shared URLs list
-- `documents_section.dart` — shared files/documents list
-- `locations_section.dart` — shared location messages list
+**Shared Content (`sections/`)**
+- `sections/media/` — images & videos grid + inline type selector
+- `sections/links/` — shared URLs list + search helper
+- `sections/documents/` — shared files list + search helper
+- `sections/locations/` — shared location messages list
