+++ lib/app/layouts/conversation_details/conversation_details.dart
@@ -3,10 +3,10 @@ import 'package:flutter_acrylic/flutter_acrylic.dart';
 import 'package:bluebubbles/app/layouts/conversation_details/widgets/chat_info.dart';
 import 'package:bluebubbles/app/state/chat_state_scope.dart';
 import 'package:bluebubbles/app/layouts/conversation_details/widgets/chat_options.dart';
-import 'package:bluebubbles/app/layouts/conversation_details/widgets/documents_section.dart';
-import 'package:bluebubbles/app/layouts/conversation_details/widgets/links_section.dart';
-import 'package:bluebubbles/app/layouts/conversation_details/widgets/locations_section.dart';
-import 'package:bluebubbles/app/layouts/conversation_details/widgets/media_grid_section.dart';
+import 'package:bluebubbles/app/layouts/conversation_details/widgets/sections/documents/documents_section.dart';
+import 'package:bluebubbles/app/layouts/conversation_details/widgets/sections/links/links_section.dart';
+import 'package:bluebubbles/app/layouts/conversation_details/widgets/sections/locations/locations_section.dart';
+import 'package:bluebubbles/app/layouts/conversation_details/widgets/sections/media/media_grid_section.dart';
 import 'package:bluebubbles/app/layouts/conversation_details/widgets/participants_list.dart';
 import 'package:bluebubbles/helpers/helpers.dart';
 import 'package:bluebubbles/app/layouts/settings/widgets/settings_widgets.dart';
