+++ lib/app/layouts/conversation_details/widgets/CLAUDE.md
@@ -2,7 +2,7 @@
 
 Reusable widgets composing the conversation details / info panel.
 
-## Files
+## Root widgets
 
 | File | Purpose |
 |------|---------|
@@ -10,12 +10,27 @@ Reusable widgets composing the conversation details / info panel.
 | `chat_options.dart` | Action row: mute, pin, archive, block, delete |
 | `contact_tile.dart` | Single participant row (avatar, name, address, remove button) |
 | `participants_list.dart` | Scrollable list of `ContactTile`s for group chats |
-| `attachments_loader.dart` | Loads and paginates shared media/attachments for the media gallery |
-| `media_gallery_card.dart` | Tappable thumbnail card for a single shared media item |
-| `media_grid_section.dart` | Grid layout section grouping media thumbnails by type |
-| `documents_section.dart` | List of shared documents (PDFs, files) |
-| `links_section.dart` | List of shared URL link previews |
-| `locations_section.dart` | List of shared location messages |
+| `attachment_section_header.dart` | Section label + "Show more" for attachment previews |
+| `attachments_loader.dart` | Loads shared attachments for media/docs/locations |
+| `media_gallery_card.dart` | Tappable thumbnail card for media or file items |
+
+## `filters/`
+
+| File | Purpose |
+|------|---------|
+| `media_filters_sheet.dart` | Shared attachment filters bottom sheet + app bar tune button |
+
+## `sections/`
+
+| Path | Purpose |
+|------|---------|
+| `media/media_grid_section.dart` | Photo/video grid (preview + full page) |
+| `media/media_filter_selector.dart` | Inline All/Images/Videos segmented control |
+| `links/links_section.dart` | Shared URL link previews |
+| `links/links_search_helper.dart` | Link search scoring and sort |
+| `documents/documents_section.dart` | Shared files/documents grid |
+| `documents/documents_search_helper.dart` | File search scoring and sort |
+| `locations/locations_section.dart` | Shared location message cards |
 
 ## Related
 - Parent panel: `../CLAUDE.md` (conversation_details)
