+++ lib/app/layouts/conversation_details/widgets/sections/documents/documents_search_helper.dart
